Starting work on some image location resolution changes
Non absolute image URIs were causing some problems with my project
structure. Beginning some changes that will allow you to have a
more diverse source structure. Starting by allowing output.dir to
be used instead of the base dir for evaluating where images should
be loaded from. This requires the images be present in the output
dir, of course. This is kludgy and will need to change in the
future, but it's a start.