Adding a way to customize how objects are chosen for use in a mediaobject
Previously the stylesheets would select the first valid element within a
mediaobject as the element to render. However, users (myself included)
may want to have a more detailed selection process when choosing among
multiple options.

Identified the namespace `usrfn` for user extension functions of this
type. Created function `usrfn:is-acceptable-mediaobject` which by
default always returns 1. This function is evaluated before any of
the built-in conditions of the stylesheet, and if it returns 0 then
`f:is-acceptable-mediaobject` will also return 0.