URI encoding the output.dir parameter when it is used for images
The path defined by output.dir may not be
a standard URI string. When using output.dir in any function that
expects a proper URI, be sure to properly prepare the input.

It might be safer and butter practice to put the iri-to-uri call
inside the functions that require URIs. For now I will assume that
the caller is going to be pleasant.