Starting work on implementing glossary.sort parameter
For the moment, just replicating the gist of the implementation
from the original docbook xsl stylesheets.

Also learned some things about the new stylesheets, like the fact
that the glossary generation _CAN_ take place in 50-normalize.xsl.
Not sure that I totally like that, but we'll see what we do about
that.