Working on recursive inclusion of glossterms from external glossary
A current limitation of the stylesheets is that if you have glossary
entries in an external stylesheet that reference other glossary entries
not in the source document, those non-source-included glossentries will
not be autogenerated.

These changes perform a second pass which catches those extra references
by recursively iterating through the glossentries that were selected
from the source document.

This is crude and incomplete (no/not good support for glossdiv), but
it works for at least a few usecases.