Changing the index generation to respect generate.index
The documentation for generate.index suggests that it can be used
to forbid autogeneration of an index, but that was not true. Now
we perform a check to make sure that, even if the index element is
empty, generate.toc is not set to false.

Addresses https://github.com/rwdalpe/xslt20-stylesheets/issues/2